Workaround for gurobi dependency
The gurobi package isn't on CRAN, you get it from the gurobi
people w/ the gurobi app itself.  renv doesn't like this. In this
workaround, I replace the version of the app stated in the lockfile
with the version of the app now shipping with the gurobi app. After
doing this, I was able to install it using
  renv::install("/Library/gurobi1102/macos_universal2/R/gurobi_11.0-2_R_4.4.0.tgz")
(By itself `renv::restore()` still didn't work.)  However,
after that renv::restore() still bombs. Grr.